Refaktorisanje JS koda u vise fajlova da kod bude citljiv generalno i svima olaksa zivot
