+++ index.html
@@ -86,6 +86,6 @@
     <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.5.2/dist/umd/popper.min.js"></script>
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
-    <script src="frontend/js/script.js"></script>
+    <script type = "module" src="frontend/js/app.js"></script>
 </body>
 </html>
